ye boi

+++ src/main/kotlin/com/example/books/persistence/PublisherRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PublisherRepository: CrudRepository<Publisher, Long> {
+interface PublisherRepository: CrudRepository<Publisher, Int> {
 }
